Update Photoframe.py
Changed image cache to use up to 90% of available disk space on the SD card... This will maximize the cache while minimizing writes to the SD card, and will free up the memory that was being used by the tmpfs.   Also added a cache clean up function... Any email that has been deleted, that image will be deleted from the cache.