Update Photoframe
Moved the wake/sleep functionality to its own function, and changed the loop type for the main loop so we catch new messages sooner.   Using the old method, it would have to complete iteration through the current messages, then on the next iteration, it would catch them at the end... so it could be as much as nearly two full passes before a new message was picked up... this new method, new messages are tacked to the end every 10 minutes, and it will pick them up on the current loop.