Update Photoframe
Improved the IMAP connection error handling
changed how images are processed and converted to minimize writes to the SD card, now initial download, conversion, and annotation is done on /tmp which can be mounted as a tmpfs   only the final result is written to the image_cache/ directory on the SD card.   All images are converted to .jpg to minimize size.